Step 1 of adding a proper numeric tower.
Depend on num-* libraries for that. They look stable and
are Apache2/MIT licensed. The main drawback is that some
of the numeric types are huge, and even bigints are 32
bytes long, which is a fair bit. I've boxed up all types
larger than this, including the behemoth 128-byte complex
rational.

For reference (in bytes):

big int 32
big rational 64
ratio of int64 16
complex of big rational 128
complex of f64 16