Fix an issue with prefixes
#### Ramblings

I realize now that I was foolish. Besides the float special
values, which should probably be represented as special
tokens, a value can be read in under float notation, but
be specified to be exact, in which case the current
implementation fails as precision has already been lost.

This could allow us to remove the notion of exactness from
the reader too, with values being returned from the lexer
as a type that matches their exactness.