Evaluate macro constructors with local bindings
When a let-syntax form is encountered, the lambda defining the macro
used to be evaluated in the global environment. This made sense because
at that point, even if the let-syntax is deeply nested in other
environments, none of the values actually exist. Unfortunately, this
meant that a macro invoked within that macro expansion would not see
the local environment at all (specifically, at expansion time, mac-env
would be set to the global environment, not the actual nested
environment).

In this change, a series of nested dummy activation frames are created
and the macro is run with these frames loaded.

We could avoid creating the frames, but then if the lambda
dereferences its environment, it would result in an illegal access.

Some other simplifications.