Fix arity checking
This had been broken since cc897c89acd18c6811179c8cf6c717a8aeaa0260
due to careless handling of the program counter when loading a procedure.

The PC was set to 0 (the first instruction in the code block), but
subsequently incremented in the VM loop, thus skipping the arity check
(always the first instruction).